requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==1.5.3
numpy==1.24.4
xgboost==1.7.5
scikit-learn==1.2.2
joblib==1.2.0
python-dotenv==1.0.0
requests==2.28.2
gunicorn==20.1.0
Werkzeug==2.2.3
tensorflow-cpu==2.10.1